Clean up error stack traces
