+++ src/error-manager/index.js
@@ -2,60 +2,15 @@
 
 const { app } = require('electron')
 const log = require('electron-log')
-const os = require('os')
 
 const isDevEnv = process.env.NODE_ENV === 'development'
 
+const getErrorDescription = require('./get-error-description')
 const showModalDialog = require('./show-modal-dialog')
 const renderMarkdownTemplate = require('./render-markdown-template')
 const openNewGithubIssue = require('./open-new-github-issue')
 const collectLogs = require('./collect-logs')
 const getDebugInfo = require('../helpers/get-debug-info')
-
-const _getErrorDescription = (params) => {
-  const { error } = { ...params }
-
-  const _title = '[BUG REPORT]'
-  const _description = 'Bug report'
-
-  if (
-    error &&
-    typeof error === 'string'
-  ) {
-    return {
-      title: _title,
-      description: [
-        'An error occurred',
-        '',
-        '```vim',
-        error,
-        '```'
-      ].join(os.EOL)
-    }
-  }
-  if (error instanceof Error) {
-    const errStr = error.toString()
-    const stack = error.stack
-      ? error.stack
-      : errStr
-
-    return {
-      title: `${_title} ${errStr}`,
-      description: [
-        'An error occurred',
-        '',
-        '```vim',
-        stack,
-        '```'
-      ].join(os.EOL)
-    }
-  }
-
-  return {
-    title: _title,
-    description: _description
-  }
-}
 
 const manageNewGithubIssue = async (params) => {
   try {
@@ -65,7 +20,7 @@ const manageNewGithubIssue = async (params) => {
     const {
       title,
       description
-    } = _getErrorDescription(params)
+    } = getErrorDescription(params)
 
     const mdIssue = renderMarkdownTemplate({
       description,
